\name{RcmdrPlugin.SensoMineR-internal}
\title{Internal RcmdrPlugin.SensoMineR objects}
\alias{defmacro}
\alias{SensArdi}
\alias{SensAverage}
\alias{SensBoxprod}
\alias{SensCarto}
\alias{SensCPA}
\alias{SensDecat}
\alias{SensDrawnapp}
\alias{SensFAST}
\alias{SensGraphinter}
\alias{Senshist}
\alias{SensIndscal}
\alias{SensOptDesign}
\alias{SensPaneliperf}
\alias{SensPanellipse}
\alias{SensPanellipse.Sess}
\alias{SensPanelmatch}
\alias{SensPanelperf}
\alias{SensPmfa}
\alias{SensScaleby}
\alias{SensTriang}
\alias{SensTriangdesign}
\description{Internal RcmdrPlugin.SensoMineR objects.}
\details{These are not to be called by the user.}
\keyword{internal}
